requirements.txt
streamlit
torch
torchvision
tensorflow
pillow
numpy
